test-gcf-helloworld/requirements.txt
google-cloud-storage==2.1.0
Pillow==9.0.0
numpy==1.22.1
